eolia-backend/src/functions/contracts/requirements.txt
fpdf2>=2.7.0
boto3>=1.28.0
